+++ extensions/nba-game-viewer/src/components/Game.tsx
@@ -15,6 +15,7 @@ const GameComponent = ({ game }: PropTypes) => {
       accessories={generateGameAccessories(game)}
       actions={
         <ActionPanel>
+          {/* eslint-disable-next-line @raycast/prefer-title-case */}
           <Action.OpenInBrowser url={game.stream} title="View Game Details on ESPN" icon={Icon.Globe} />
         </ActionPanel>
       }