+++ extensions/nba-game-viewer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # NBA Game Viewer Changelog
 
+## [Fixes] - 2024-08-27
+
+- If `useLastValue` is not selected in preference, then default values for dropdown is picked from preference.
+- Uses bundled heart icon from raycast instead of a custom one.
+
 ## [Improvements] - 2024-08-22
 
 - Adds the ability to select either the NBA or WNBA league to view games, standings, and more.